requirements.txt
pillow
pynput
scapy
